requirements.txt
# FastAPI & ASGI
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Async DB support
SQLAlchemy==1.4.49
asyncpg==0.29.0
databases==0.8.0


# Alembic for migrations
alembic==1.13.1

# Security
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1

# Email/SMS utils
httpx==0.27.0

# Logging
loguru==0.7.2

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5

# Multipart/form handling
python-multipart==0.0.9
